refactor(broker): split aggregator.rs into directory layout (C31 of PR-3)
Pure mechanical move plus relocation of three Batch-related types from
lib.rs into the aggregator module where they belong.

aggregator.rs (1715 lines) becomes:

- aggregator/mod.rs (~32 lines): module decls + pub use service::AggregatorService
  and pub use types::{AggregationState, Batch, BatchStatus}.
- aggregator/error.rs (~32 lines): AggregatorErr enum + coded_error_impl!.
- aggregator/types.rs (~75 lines): public batch types Batch / BatchStatus /
  AggregationState, moved from lib.rs; plus the pub(crate)
  AggregateProofsResult helper that was internal in aggregator.rs (its
  fields are now pub(crate) so service.rs can construct it across modules).
- aggregator/service.rs (~1690 lines): AggregatorService struct, impl, the
  BrokerService impl, and the original test module.

External path crate::aggregator::AggregatorService::new preserved via the
service re-export. crate::Batch / crate::BatchStatus / crate::AggregationState
preserved via pub use aggregator::{...} in lib.rs. The corresponding type
definitions and four imports (alloy::U256, chrono::{DateTime, Utc},
risc0_zkvm::sha::Digest, serde::{Deserialize, Serialize}) are removed
from lib.rs since nothing else there used them.

AggregatorErr is no longer re-exported from aggregator/mod.rs (nothing
outside the module imports it, and pub use error::AggregatorErr was flagged
as an unused import).
