+++ crates/broker/src/lib.rs
@@ -12,17 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use alloy::primitives::U256;
 use boundless_market::contracts::ProofRequest;
-use chrono::{DateTime, Utc};
 pub use config::Config;
 pub use config::ConfigLock;
-use risc0_zkvm::sha::Digest;
 pub use rpc_retry_policy::CustomRetryPolicy;
-use serde::{Deserialize, Serialize};
 pub(crate) use storage::ConfigurableDownloader;
 
 pub(crate) mod aggregator;
+pub use aggregator::{AggregationState, Batch, BatchStatus};
 pub mod args;
 pub(crate) mod block_history;
 mod broker;
@@ -64,52 +61,6 @@ pub use boundless_market::prover_utils::{
 };
 pub(crate) use order_types::proving_order_from_request;
 pub use order_types::{CompressionType, Order, OrderStateChange, OrderStatus};
-
-#[derive(sqlx::Type, Default, Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub enum BatchStatus {
-    #[default]
-    Aggregating,
-    PendingCompression,
-    Complete,
-    PendingSubmission,
-    Submitted,
-    Failed,
-}
-
-#[derive(Serialize, Deserialize, Clone)]
-pub struct AggregationState {
-    pub guest_state: risc0_aggregation::GuestState,
-    /// All claim digests in this aggregation.
-    /// This collection can be used to construct the aggregation Merkle tree and Merkle paths.
-    pub claim_digests: Vec<Digest>,
-    /// Proof ID for the STARK proof that compresses the root of the aggregation tree.
-    pub proof_id: String,
-    /// Proof ID for the Groth16 proof that compresses the root of the aggregation tree.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub groth16_proof_id: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Default, Clone)]
-pub struct Batch {
-    pub status: BatchStatus,
-    /// Orders from the market that are included in this batch.
-    pub orders: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub assessor_proof_id: Option<String>,
-    /// Tuple of the current aggregation state, as committed by the set builder guest, and the
-    /// proof ID for the receipt that attests to the correctness of this state.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub aggregation_state: Option<AggregationState>,
-    /// When the batch was initially created.
-    pub start_time: DateTime<Utc>,
-    /// The deadline for the batch, which is the earliest deadline for any order in the batch.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub deadline: Option<u64>,
-    /// The total fees for the batch, which is the sum of fees from all orders.
-    pub fees: U256,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error_msg: Option<String>,
-}
 
 #[cfg(feature = "test-utils")]
 pub mod test_utils;
